Move shared search stuff to Query namespace
