+++ src/Extend/SimpleFlarumSearch.php
@@ -61,7 +61,7 @@ class SimpleFlarumSearch implements ExtenderInterface
      *
      * The callback can be a closure or an invokable class, and should accept:
      * - Flarum\Search\SearchState $search
-     * - Flarum\Search\SearchCriteria $criteria
+     * - Flarum\Query\QueryCriteria $criteria
      */
     public function addSearchMutator($callback)
     {
